requirements.txt
cobra>=0.29.0
pandas>=2.0
numpy>=1.24
matplotlib>=3.8
pyyaml>=6.0
networkx>=3.0
scipy>=1.11
pypdf>=4.0
